Add a note about Magento's introduction of a progressive web app (PWA)
Recently, Magento announced plans to publish a mechanism for developers
to, in turn, develop and publish PWAs. At this stage, details are
limited to just the dev blog post, however it's an interesting area and
in the authors opinion, deserves to be tracked as part of the roadmap.

Design Notes:

unspecified:

  It's not clear whether the PWA studio will be released with Magento,
  or as a separate product.